+++ src/layer/overlay-layer-group.ts
@@ -1,7 +1,10 @@
-import { Subject } from 'rxjs';
 import OlMap from 'ol/Map.js';
 import { getOlcUid, olcUidKey } from '../uid.js';
-import { LayerGroup, type LayerGroupOptions } from './layer-group.js';
+import {
+  LayerGroup,
+  type LayerGroupOnSignature,
+  type LayerGroupOptions,
+} from './layer-group.js';
 import {
   createEmpty as olCreateEmptyExtent,
   extend as olExtend,
@@ -15,27 +18,74 @@ import { Geometry as OlGeometry } from 'ol/geom.js';
 import OlCollection from 'ol/Collection.js';
 import OlSourceCluster from 'ol/source/Cluster.js';
 import { getFeaturesExtent } from '../feature/utils.js';
+import type { EventsKey } from 'ol/events.js';
+import BaseEvent from 'ol/events/Event.js';
 
 export const DefaultOverlayLayerGroupUid = 'olcOverlayLayerGroupUid';
 
-/**
- * Event definition for "affected" event of a feature in the layerGroup.
- */
-export interface FeatureAffected {
-  [olcUidKey]: string;
-  reason: string;
-  affected: OlFeature[];
-  noMoreAffected?: OlFeature[];
-}
+// Event types names.
+export const FeatureAffectedEventType = 'olcFeatureAffected';
+export const FeaturePropertyChangedEventType = 'olcFeaturePropertyChanged';
 
 /**
- * Event definition for property change in a feature of a layer in the layerGroup.
+ * Event definition for "affected" event of a feature in the layerGroup.
+ * To provide changed-like-event, without touching the original features.
+ * With source layer uid, free reason why/by/for it's fired, the affected features
+ * and the not anymore affected features.
+ * Example: fire/listen to selected/deselected (as reason) features without touching
+ * the real features and having to listen to every layer individually.
  */
-export interface FeaturePropertyChanged {
-  [olcUidKey]: string;
-  propertyKey: string;
-  features: OlFeature[];
+export class FeatureAffectedEvent extends BaseEvent {
+  static readonly type = FeatureAffectedEventType;
+  readonly [olcUidKey]: string;
+  readonly reason: string;
+  readonly affected: OlFeature[];
+  readonly noMoreAffected?: OlFeature[];
+
+  constructor(
+    layerUid: string,
+    reason: string,
+    affected: OlFeature[],
+    noMoreAffected?: OlFeature[],
+  ) {
+    super(FeatureAffectedEvent.type);
+    this[olcUidKey] = layerUid;
+    this.reason = reason;
+    this.affected = affected;
+    this.noMoreAffected = noMoreAffected;
+  }
 }
+
+/*
+ * An event to listen on feature property changes in layers of this group and without the need of having
+ * the feature itself.
+ */
+export class FeaturePropertyChangedEvent extends BaseEvent {
+  static readonly type = FeaturePropertyChangedEventType;
+  readonly [olcUidKey]: string;
+  readonly propertyKey: string;
+  readonly features: OlFeature[];
+
+  constructor(layerUid: string, propertyKey: string, features: OlFeature[]) {
+    super(FeaturePropertyChangedEvent.type);
+    this[olcUidKey] = layerUid;
+    this.propertyKey = propertyKey;
+    this.features = features;
+  }
+}
+
+// Typesafe ol events.
+export type FeatureAffectedOnSignature = (
+  type: typeof FeatureAffectedEventType,
+  listener: (event: FeatureAffectedEvent) => void,
+) => EventsKey;
+export type FeaturePropertyChangedOnSignature = (
+  type: typeof FeaturePropertyChangedEventType,
+  listener: (event: FeaturePropertyChangedEvent) => void,
+) => EventsKey;
+export type OverlayLayerGroupOnSignature = FeatureAffectedOnSignature &
+  FeaturePropertyChangedOnSignature &
+  LayerGroupOnSignature; // default from parent class.
 
 /**
  * LayerGroup specialized to manage layers with features (mostly vector layers).
@@ -43,27 +93,15 @@ export interface FeaturePropertyChanged {
  * The default position is 20.
  */
 export class OverlayLayerGroup extends LayerGroup {
-  /**
-   * To provide changed-like-event, without touching the original features.
-   * With source layer uid, free reason why/by/for it's emitted, the affected features
-   * and the not anymore affected features.
-   * Example: emit/listen to selected/deselected (as reason) features without touching
-   * the real features and having to listen to every layer individually.
-   */
-  readonly featuresAffected: Subject<FeatureAffected>;
-  /**
-   * Event to observe feature property change in layers of this group and without the need of having
-   * the feature itself.
-   */
-  readonly featuresPropertyChanged: Subject<FeaturePropertyChanged>;
+  declare on: OverlayLayerGroupOnSignature;
+  declare once: OverlayLayerGroupOnSignature;
+  declare un: OverlayLayerGroupOnSignature;
 
   constructor(map: OlMap, options: LayerGroupOptions = {}) {
     const layerGroupUid = options.groupUid ?? DefaultOverlayLayerGroupUid;
     super(map);
     const position = options.position ?? 20;
     this.addLayerGroup(layerGroupUid, position);
-    this.featuresAffected = new Subject<FeatureAffected>();
-    this.featuresPropertyChanged = new Subject<FeaturePropertyChanged>();
   }
 
   /**
@@ -194,27 +232,24 @@ export class OverlayLayerGroup extends LayerGroup {
   }
 
   /**
-   * Emit an "affected" feature event.
+   * Fires an "affected" feature event.
    */
-  emitFeaturesAffected(
+  dispatchFeaturesAffected(
     layerUid: string,
     reason: string,
     affected: OlFeature<OlGeometry>[],
     noMoreAffected?: OlFeature<OlGeometry>[],
   ) {
-    this.featuresAffected.next({
-      [olcUidKey]: layerUid,
-      reason,
-      affected,
-      noMoreAffected,
-    });
+    this.dispatchEvent(
+      new FeatureAffectedEvent(layerUid, reason, affected, noMoreAffected),
+    );
   }
 
   /**
    * Set a property of every given feature with the same value.
    * If you want to set different values, set the feature manually (with silent=true), then call
-   * emitFeaturePropertyChange to redraw the ol layer and emit a featuresPropertyChanged event.
-   * Or call featuresPropertyChanged manually if necessary.
+   * dispatchFeaturePropertyChanged to redraw the ol layer and fires a featuresPropertyChanged event.
+   * Or call featuresPropertyChange manually if necessary.
    */
   setFeaturesProperty(
     layerUid: string,
@@ -225,23 +260,19 @@ export class OverlayLayerGroup extends LayerGroup {
     features.forEach((feature) => {
       feature.set(propertyKey, value, true);
     });
-    this.emitFeaturePropertyChanged(layerUid, features, propertyKey);
+    this.dispatchFeaturePropertyChanged(layerUid, features, propertyKey);
   }
 
   /**
    * Emits a "featuresPropertyChanged" and call a "changed" event on the matching OL layer.
    */
-  emitFeaturePropertyChanged(
+  dispatchFeaturePropertyChanged(
     layerUid: string,
     features: OlFeature[],
     propertyKey: string,
   ) {
     this.getLayer(layerUid)?.changed();
-    this.featuresPropertyChanged.next({
-      [olcUidKey]: layerUid,
-      propertyKey,
-      features,
-    });
+    this.dispatchEvent(new FeaturePropertyChangedEvent(layerUid, propertyKey, features));
   }
 
   /**